pyproject.toml
[project]
name = "od-prank-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "alembic>=1.16.4",
    "celery>=5.4.0",
    "deprecated>=1.2.18",
    "elevenlabs>=2.10.0",
    "fastapi>=0.116.1",
    "greenlet>=3.2.4",
    "langchain-core>=0.3.74",
    "langchain-openai>=0.3.30",
    "langgraph>=0.6.5",
    "numpy>=2.3.2",
    "pydantic-settings>=2.10.1",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "redis>=6.4.0",
    "sqlalchemy[asyncio]>=2.0.43",
    "sqlalchemy-continuum>=1.4.2",
    "structlog>=25.4.0",
    "supabase>=2.18.1",
    "telnyx>=2.1.6",
    "uvicorn>=0.35.0",
    "phonenumbers>=8.13.48",
    "gunicorn>=23.0.0",
    "uvicorn-worker>=0.3.0",
    "stripe>=12.5.1",
    "pytz>=2025.2",
    "psycopg[binary]>=3.1.20",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.10",
]
